fix: types

+++ tasks/gaivoronskiy_m_average_vector_sum/mpi/src/ops_mpi.cpp
@@ -34,8 +34,8 @@ bool GaivoronskiyMAverageVecSumMPI::PreProcessingImpl() {
     distributed_values_.clear();
   }
 
-  int size_to_share = static_cast<MPI_INT>(total_size_);
-  MPI_Bcast(&size_to_share, 1, MPI_UNSIGNED_LONG, 0, MPI_COMM_WORLD);
+  int size_to_share = static_cast<int>(total_size_);
+  MPI_Bcast(&size_to_share, 1, MPI_INT, 0, MPI_COMM_WORLD);
   total_size_ = static_cast<std::size_t>(size_to_share);
 
   local_sum_ = 0.0;
